Fixed buffer merge with parent present on DEMOLISH
If no parent was defined, an exception was thrown, but DEMOLISH is run
inside a try/catch block so we didn't see it on normal situations.

Still, this code is safer. The test should always work fine, unless
we start seeing exceptions from DEMOLISH methods *and* the bug is
re-introduced.

Signed-off-by: Pedro Melo <melo@simplicidade.org>