Revamp setup() logic:
All widgets setup entry point is now Ferdinand->setup(), which takes
the class name to setup and a the meta argument.

This code enforces the call to setup_check().

Added setup_map() to build a Ferdinand::Map. This method is deprecated, it
will disapear, when we convert Ferdinand::Map to a widget.

Signed-off-by: Pedro Melo <melo@simplicidade.org>