Make _process_buttons in ListButtons extensible:
During the setup_args phase, classes using this role should add the button
name to the $args->{btns_to_check}.

If you require a prefix-match button, push the button name as a scalarref.

The _process_buttons() code will check all btns in the list btns_to_check.
It will use the "${btn_name}_id" as the parameter name, and
"_${btn_name}_action" as the method to invoke if the button is pressed.

Each button _action method is called with:

 * $ctx: the current Ferdinand context;
 * $elems: the current list of elements;
 * $id_map: a hashref with IDs of the elements so far;
 * $btn_arg: on prefix buttons, the argument.

The _process_buttons() returns:

 * $elems: the elements for this list;
 * $id_map: a HashRef with all the element IDs we've seen;
 * $btns_used: a HashRef with the button pressed, if any.

Signed-off-by: Pedro Melo <melo@simplicidade.org>